# api_collection/user_profile_management_nest.openapi.json
{"openapi":"3.1.0","info":{"title":"user_profile_management_nest","description":"","version":"1.0.0"},"tags":[{"name":"Organization"},{"name":"User"}],"paths":{"/organization/create":{"post":{"summary":"Create Organization","deprecated":false,"description":"","tags":["Organization"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"name":"{% faker company.name %}","contact_email":"{% faker internet.exampleEmail %}","address":[{"type":"PRIMARY","label":"Home","address":"{% faker address.streetAddress %}","city":"{% faker address.city %}","state":"{% faker address.state %}","country":"{% faker address.country %}","country_code":"{% faker address.countryCode %}","zipcode":"{% faker address.zipCode %}","latitude":"{% faker address.latitude %}","longitude":"{% faker address.longitude %}","offset":"+9:30","zone":"{% faker address.timeZone %}","is_default":true},{"type":"SECONDARY","label":"Apartment","address":"{% faker address.streetAddress %}","city":"{% faker address.city %}","state":"{% faker address.state %}","country":"{% faker address.country %}","country_code":"{% faker address.countryCode %}","zipcode":"{% faker address.zipCode %}","latitude":"{% faker address.latitude %}","longitude":"{% faker address.longitude %}","offset":"+9:30","zone":"{% faker address.timeZone %}","is_default":false}],"contact":[{"type":"PRIMARY","label":"Home","country_code":"+01","number":"{% faker phone.number %}","is_default":true},{"type":"SECONDARY","label":"Apartment","country_code":"+01","number":"{% faker phone.number %}","is_default":false}],"social_profiles":[{"label":"Facebook","link":"{% faker internet.url %}"},{"label":"Instagream","link":"{% faker internet.url %}"},{"label":"Twitter","link":"{% faker internet.url %}"},{"label":"LinkedIn","link":"{% faker internet.url %}"}],"logo":"{% faker image.imageUrl %}"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/organization/list":{"get":{"summary":"Get Organization List","deprecated":false,"description":"","tags":["Organization"],"parameters":[{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"limit","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}},{"name":"search","in":"query","description":"","required":false,"example":"abcd","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/organization/fd27bc65-1a8d-4f6c-a22b-6de95a52a1c4":{"get":{"summary":"Get Organization Details","deprecated":false,"description":"","tags":["Organization"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]},"patch":{"summary":"Update Organization","deprecated":false,"description":"","tags":["Organization"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"name":"{% faker company.name %}","contact_email":"{% faker internet.exampleEmail %}","address":[{"type":"PRIMARY","label":"Home","address":"{% faker address.streetAddress %}","city":"{% faker address.city %}","state":"{% faker address.state %}","country":"{% faker address.country %}","country_code":"{% faker address.countryCode %}","zipcode":"{% faker address.zipCode %}","latitude":"{% faker address.latitude %}","longitude":"{% faker address.longitude %}","offset":"+9:30","zone":"{% faker address.timeZone %}","is_default":true},{"type":"SECONDARY","label":"Apartment","address":"{% faker address.streetAddress %}","city":"{% faker address.city %}","state":"{% faker address.state %}","country":"{% faker address.country %}","country_code":"{% faker address.countryCode %}","zipcode":"{% faker address.zipCode %}","latitude":"{% faker address.latitude %}","longitude":"{% faker address.longitude %}","offset":"+9:30","zone":"{% faker address.timeZone %}","is_default":false}],"contact":[{"type":"PRIMARY","label":"Home","country_code":"+01","number":"{% faker phone.number %}","is_default":true},{"type":"SECONDARY","label":"Apartment","country_code":"+01","number":"{% faker phone.number %}","is_default":false}],"social_profiles":[{"label":"Facebook","link":"{% faker internet.url %}"},{"label":"Instagream","link":"{% faker internet.url %}"},{"label":"Twitter","link":"{% faker internet.url %}"},{"label":"LinkedIn","link":"{% faker internet.url %}"}],"logo":"{% faker image.imageUrl %}"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]},"delete":{"summary":"Delete Organization (Soft)","deprecated":false,"description":"","tags":["Organization"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/create":{"post":{"summary":"Create User","deprecated":false,"description":"","tags":["User"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":"{\n    \"name_prefix\": \"{% faker name.prefix %}\",\n    \"first_name\": \"{% faker name.firstName %}\",\n    \"middle_name\": \"{% faker name.middleName %}\",\n    \"last_name\": \"{% faker name.lastName %}\",\n    \"name_suffix\": \"{% faker name.suffix %}\",\n    \"email\": \"{% faker internet.email %}\",\n    \"username\": \"{% faker internet.userName %}\",\n    // \"organization_id\": \"{% faker datatype.uuid %}\",\n    \"password\": \"Demo@12345\",\n    \"identification\": [\n        {\n            \"type\": \"SSN\",\n            \"value\": \"{% faker finance.creditCardNumber %}\"\n        }\n    ],\n    \"address\": [\n        {\n            \"type\": \"PRIMARY\",\n            \"label\": \"Home\",\n            \"address\": \"{% faker address.streetAddress %}\",\n            \"city\": \"{% faker address.city %}\",\n            \"state\": \"{% faker address.state %}\",\n            \"country\": \"{% faker address.country %}\",\n            \"country_code\": \"{% faker address.countryCode %}\",\n            \"zipcode\": \"{% faker address.zipCode %}\",\n            \"latitude\": \"{% faker address.latitude %}\",\n            \"longitude\": \"{% faker address.longitude %}\",\n            \"offset\": \"+9:30\",\n            \"zone\": \"{% faker address.timeZone %}\",\n            \"is_default\": true\n        },\n        {\n            \"type\": \"SECONDARY\",\n            \"label\": \"Apartment\",\n            \"address\": \"{% faker address.streetAddress %}\",\n            \"city\": \"{% faker address.city %}\",\n            \"state\": \"{% faker address.state %}\",\n            \"country\": \"{% faker address.country %}\",\n            \"country_code\": \"{% faker address.countryCode %}\",\n            \"zipcode\": \"{% faker address.zipCode %}\",\n            \"latitude\": \"{% faker address.latitude %}\",\n            \"longitude\": \"{% faker address.longitude %}\",\n            \"offset\": \"+9:30\",\n            \"zone\": \"{% faker address.timeZone %}\",\n            \"is_default\": false\n        }\n    ],\n    \"contact\": [\n        {\n            \"type\": \"PRIMARY\",\n            \"label\": \"Home\",\n            \"country_code\": \"+01\",\n            \"number\": \"{% faker phone.number %}\",\n            \"is_default\": true\n        },\n        {\n            \"type\": \"SECONDARY\",\n            \"label\": \"Apartment\",\n            \"country_code\": \"+01\",\n            \"number\": \"{% faker phone.number %}\",\n            \"is_default\": false\n        }\n    ],\n    \"social_profiles\": [\n        {\n            \"label\": \"Facebook\",\n            \"link\": \"{% faker internet.url %}\"\n        },\n        {\n            \"label\": \"Instagream\",\n            \"link\": \"{% faker internet.url %}\"\n        },\n        {\n            \"label\": \"Twitter\",\n            \"link\": \"{% faker internet.url %}\"\n        },\n        {\n            \"label\": \"LinkedIn\",\n            \"link\": \"{% faker internet.url %}\"\n        }\n    ],\n    \"avatar\": \"{% faker image.imageUrl %}\",\n    \"meta_data\": {\n        \"gender\": \"{% faker name.gender %}\",\n        \"dob\": \"1989-11-27 18:04:06\",\n        \"theme_code\": \"DEFAULT\",\n        \"is_super_admin\": true\n    }\n}"}}},"responses":{"201":{"description":"","content":{"*/*":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/list":{"get":{"summary":"Get User List","deprecated":false,"description":"","tags":["User"],"parameters":[{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"limit","in":"query","description":"","required":false,"example":"10","schema":{"type":"string"}},{"name":"search","in":"query","description":"","required":false,"example":"abcd","schema":{"type":"string"}}],"responses":{"400":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"array","items":{"type":"string"}},"error":{"type":"string"},"statusCode":{"type":"integer"}},"required":["message","error","statusCode"]}}},"headers":{}}},"security":[]}},"/user/register":{"post":{"summary":"Register User","deprecated":false,"description":"","tags":["User"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"name_prefix":"{% faker name.prefix %}","first_name":"{% faker name.firstName %}","middle_name":"{% faker name.middleName %}","last_name":"{% faker name.lastName %}","name_suffix":"{% faker name.suffix %}","email":"{% faker internet.email %}","password":"{{'Demo@1234'}}"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/user/c2a7aa98-677f-49eb-ab31-8ba85cd97cef":{"get":{"summary":"Get User Details","deprecated":false,"description":"","tags":["User"],"parameters":[],"responses":{"400":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"},"error":{"type":"string"},"statusCode":{"type":"integer"}},"required":["message","error","statusCode"]}}},"headers":{}}},"security":[]}},"/user/c8b4e6fd-193c-4e8d-9f26-d31e82f28c35":{"patch":{"summary":"Update User","deprecated":false,"description":"","tags":["User"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":"{\n    // \"name_prefix\": \"{% faker name.prefix %}\",\n    // \"first_name\": \"{% faker name.firstName %}\",\n    // \"middle_name\": \"{% faker name.middleName %}\",\n    // \"last_name\": \"{% faker name.lastName %}\",\n    // \"name_suffix\": \"{% faker name.suffix %}\",\n    // \"email\": \"{% faker internet.email %}\",\n    // \"username\": \"{% faker internet.userName %}\",\n    // \"organization_id\": \"{% faker datatype.uuid %}\",\n    // \"password\": \"Demo@12345\",\n    \"identification\": [\n        {\n            \"type\": \"SSN\",\n            \"value\": \"{% faker finance.creditCardNumber %}\"\n        }\n    ],\n    // \"address\": [\n    //     {\n    //         \"type\": \"PRIMARY\",\n    //         \"label\": \"Home\",\n    //         \"address\": \"{% faker address.streetAddress %}\",\n    //         \"city\": \"{% faker address.city %}\",\n    //         \"state\": \"{% faker address.state %}\",\n    //         \"country\": \"{% faker address.country %}\",\n    //         \"country_code\": \"{% faker address.countryCode %}\",\n    //         \"zipcode\": \"{% faker address.zipCode %}\",\n    //         \"location\": {\n    //             \"latitude\": \"{% faker address.latitude %}\",\n    //             \"longitude\": \"{% faker address.longitude %}\"\n    //         },\n    //         \"timezone\": {\n    //             \"offset\": \"+9:30\",\n    //             \"zone\": \"{% faker address.timeZone %}\"\n    //         },\n    //         \"is_default\": true\n    //     },\n    //     {\n    //         \"type\": \"SECONDARY\",\n    //         \"label\": \"Apartment\",\n    //         \"address\": \"{% faker address.streetAddress %}\",\n    //         \"city\": \"{% faker address.city %}\",\n    //         \"state\": \"{% faker address.state %}\",\n    //         \"country\": \"{% faker address.country %}\",\n    //         \"country_code\": \"{% faker address.countryCode %}\",\n    //         \"zipcode\": \"{% faker address.zipCode %}\",\n    //         \"location\": {\n    //             \"latitude\": \"{% faker address.latitude %}\",\n    //             \"longitude\": \"{% faker address.longitude %}\"\n    //         },\n    //         \"timezone\": {\n    //             \"offset\": \"+9:30\",\n    //             \"zone\": \"{% faker address.timeZone %}\"\n    //         },\n    //         \"is_default\": false\n    //     }\n    // ],\n    // \"contact\": [\n    //     {\n    //         \"type\": \"PRIMARY\",\n    //         \"label\": \"Home\",\n    //         \"country_code\": \"+01\",\n    //         \"number\": \"{% faker phone.number %}\",\n    //         \"is_default\": true\n    //     },\n    //     {\n    //         \"type\": \"SECONDARY\",\n    //         \"label\": \"Apartment\",\n    //         \"country_code\": \"+01\",\n    //         \"number\": \"{% faker phone.number %}\",\n    //         \"is_default\": false\n    //     }\n    // ],\n    // \"social_profiles\": [\n    //     {\n    //         \"label\": \"Facebook\",\n    //         \"link\": \"{% faker internet.url %}\"\n    //     },\n    //     {\n    //         \"label\": \"Instagream\",\n    //         \"link\": \"{% faker internet.url %}\"\n    //     },\n    //     {\n    //         \"label\": \"Twitter\",\n    //         \"link\": \"{% faker internet.url %}\"\n    //     },\n    //     {\n    //         \"label\": \"LinkedIn\",\n    //         \"link\": \"{% faker internet.url %}\"\n    //     }\n    // ],\n    // \"avatar\": \"{% faker image.imageUrl %}\",\n    // \"meta_data\": {\n    //     \"gender\": \"{% faker name.gender %}\",\n    //     \"dob\": \"1989-11-27 18:04:06\",\n    //     \"theme_code\": \"DEFAULT\",\n    //     \"is_super_admin\": true\n    // }\n}"}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]},"delete":{"summary":"Delete User","deprecated":false,"description":"","tags":["User"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}}},"components":{"schemas":{},"securitySchemes":{}},"servers":[{"url":"http://localhost:3000","description":"Develop Env"}]}